+++ openai-java-core/src/main/kotlin/com/openai/services/async/BetaServiceAsyncImpl.kt
@@ -3,6 +3,7 @@
 package com.openai.services.async
 
 import com.openai.core.ClientOptions
+import com.openai.core.http.Headers
 import com.openai.services.async.beta.AssistantServiceAsync
 import com.openai.services.async.beta.AssistantServiceAsyncImpl
 import com.openai.services.async.beta.ThreadServiceAsync
@@ -11,9 +12,14 @@ import com.openai.services.async.beta.VectorStoreServiceAsync
 import com.openai.services.async.beta.VectorStoreServiceAsyncImpl
 
 class BetaServiceAsyncImpl
-constructor(
+internal constructor(
     private val clientOptions: ClientOptions,
 ) : BetaServiceAsync {
+
+    companion object {
+
+        private val DEFAULT_HEADERS = Headers.builder().put("OpenAI-Beta", "assistants=v2").build()
+    }
 
     private val vectorStores: VectorStoreServiceAsync by lazy {
         VectorStoreServiceAsyncImpl(clientOptions)