+++ openai-java-core/src/main/kotlin/com/openai/services/blocking/BetaServiceImpl.kt
@@ -3,6 +3,7 @@
 package com.openai.services.blocking
 
 import com.openai.core.ClientOptions
+import com.openai.core.http.Headers
 import com.openai.services.blocking.beta.AssistantService
 import com.openai.services.blocking.beta.AssistantServiceImpl
 import com.openai.services.blocking.beta.ThreadService
@@ -11,9 +12,14 @@ import com.openai.services.blocking.beta.VectorStoreService
 import com.openai.services.blocking.beta.VectorStoreServiceImpl
 
 class BetaServiceImpl
-constructor(
+internal constructor(
     private val clientOptions: ClientOptions,
 ) : BetaService {
+
+    companion object {
+
+        private val DEFAULT_HEADERS = Headers.builder().put("OpenAI-Beta", "assistants=v2").build()
+    }
 
     private val vectorStores: VectorStoreService by lazy { VectorStoreServiceImpl(clientOptions) }